requirements.txt
numpy>=1.24
opencv-python>=4.8
mediapipe==0.10.14
torch>=2.1
tqdm>=4.66
Pillow>=10.0
